Support `root.toString()`
