+++ .gitignore
@@ -1,3 +1,6 @@
+# Created by https://www.gitignore.io/api/node
+
+### Node ###
 # Logs
 logs
 *.log
@@ -57,3 +60,6 @@ typings/
 # dotenv environment variables file
 .env
 
+
+
+# End of https://www.gitignore.io/api/node
