+++ lib/parser.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const Input = require("postcss/lib/input");
-const Root = require("postcss/lib/root");
+const Document = require("./document");
 const htmlParser = require("./html-parser");
 const mdParser = require("./markdown-parser");
 const BlockFixer = require("./block-fixer");
@@ -21,7 +21,7 @@ function parser (source, opts) {
 	if (!styleTag && !stylesMd && !filenameMatch(/\.(?:md|markdown)$/i)) {
 		return;
 	}
-	const document = new Root();
+	const document = new Document();
 	const documentLocalFixer = new BlockFixer(source);
 	let index = 0;
 	[].concat(styleTag, stylesMd).filter(Boolean).sort(function (a, b) {
@@ -40,49 +40,6 @@ function parser (source, opts) {
 			line: 1,
 			column: 1,
 		},
-	};
-
-	document.each = function (callback) {
-		let wasBreak, lastResult;
-		this.nodes.forEach(node => {
-			const result = node.each(callback);
-			if (result === false) {
-				wasBreak = true;
-			} else {
-				lastResult = result;
-			}
-		});
-		if (wasBreak) {
-			return false;
-		} else {
-			return lastResult;
-		}
-	};
-
-	document.append = function () {
-		this.last.append.apply(
-			this.last,
-			Array.from(arguments)
-		);
-		return this;
-	};
-
-	document.prepend = function () {
-		this.first.prepend.apply(
-			this.first,
-			Array.from(arguments)
-		);
-		return this;
-	};
-
-	document.insertBefore = function (exist, add) {
-		exist.prepend(add);
-		return this;
-	};
-
-	document.insertAfter = function (exist, add) {
-		exist.append(add);
-		return this;
 	};
 
 	return document;
