+++ package.json
@@ -17,8 +17,8 @@
       "lcov",
       "text"
     ],
-    "cache": true,
     "all": true,
+    "cache": true,
     "check-coverage": true
   },
   "repository": {
@@ -68,7 +68,7 @@
     "eslint-plugin-promise": "^3.7.0",
     "eslint-plugin-standard": "^3.0.1",
     "mocha": "^5.0.4",
-    "nyc": "^11.4.1",
+    "nyc": "^11.6.0",
     "postcss": "^6.0.19",
     "postcss-less": "^1.1.3",
     "postcss-parser-tests": "^6.2.1",
@@ -77,7 +77,7 @@
     "postcss-scss": "^1.0.4",
     "strip-bom": "^3.0.0",
     "stylefmt": "^6.0.0",
-    "stylelint": "^9.1.1",
+    "stylelint": "^9.1.2",
     "sugarss": "^1.0.1"
   }
 }
